% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_propensities}
\alias{calc_propensities}
\title{Propensity Calculation}
\usage{
calc_propensities()
}
\value{
None
}
\description{
Calculates the propensities of all PKC model reactions and stores them in the vector amu.
}
\examples{
calc_propensities() 
}
