% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalciumModelsLibrary.R
\docType{package}
\name{CalciumModelsLibrary}
\alias{CalciumModelsLibrary}
\alias{CalciumModelsLibrary-package}
\title{CalciumModelsLibrary: A package for calcium-sensitive protein models.}
\description{
The CalciumModelsLibrary package provides different models for
calcium oscillations and a library of models for
calcium-sensitive proteins.
}
\section{CalciumModelsLibrary functions}{
 ...
}

