% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calmodulin}
\alias{calmodulin}
\title{Couple a simulated calmodulin protein to a given calcium time series.}
\usage{
calmodulin(time, calcium, dt, vol, k_on, Km, k_off, E0_conc, h)
}
\arguments{
\item{time}{A numeric vector: the times of the observations.}

\item{calcium}{A numeric vector: the concentrations of cytosolic calcium [nmol].}

\item{dt}{A numeric, the time interval between two output samples.}

\item{vol}{A numeric, the volume of the system [l].}

\item{k_on}{A numeric, the on rate of the protein.}

\item{Km}{A numeric, the half-maximal constant of the Hill activation.}

\item{k_off}{A numeric, the off rate of the protein.}

\item{E0_conc}{A numeric, the total concentration of protein [nmol].}

\item{h}{An integer, the cooperativity of the Hill activation.}
}
\value{
A dataframe with time and the active protein time series as columns.
}
\description{
Takes a calcium time series and simulates a Ca-dependent protein, e.g.
Calmodulin.
}
\details{
This is the deterministic equivalent of the protein activation process.
dE_act/dt = E_inact * (k_act * Ca_cyt^p) / (K_m^p + Ca_cyt^p) - k_inact * E_act
}
\examples{
calmodulin()
}
