% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calmodulin_calculate_amu}
\alias{calmodulin_calculate_amu}
\title{Propensity Calculation}
\usage{
calmodulin_calculate_amu()
}
\value{
None
}
\description{
Calculates the propensities of all Calmodulin model reactions and stores them in the vector amu.
}
\examples{
calculate_amu() 
}
