% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calmodulin_stM}
\alias{calmodulin_stM}
\title{Define stoichiometric matrix of the model}
\usage{
calmodulin_stM()
}
\value{
A numeric matrix: the stoichiometric matrix
}
\description{
Create and return the stoichiometric matrix of the model as a numeric matrix.
}
\examples{
calmodulin_stM()
}
