% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calmodulin_update_system}
\alias{calmodulin_update_system}
\title{System Update}
\usage{
calmodulin_update_system(rIndex)
}
\arguments{
\item{rIndex}{An unsigned integer: the id of the chosen reaction.}
}
\value{
void
}
\description{
Changes the system state (updates the particle numbers) by instantiating a chosen reaction.
}
\examples{
update_system() 
}
