% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{camkii}
\alias{camkii}
\title{Couple a simulated CamKII protein to a given calcium time series.}
\usage{
camkii(time, calcium, dt, vol, a, b, c, k_IB, k_BI, k_PT, k_TP, k_TA, k_AT,
  k_AA, c_I, c_B, c_P, c_T, c_A, camT, Kd, Vm_phos, Kd_phos, totalC, Wi_conc,
  Wb_conc, Wp_conc, Wt_conc, Wa_conc, h)
}
\arguments{
\item{time}{A numeric vector: the times of the observations.}

\item{calcium}{A numeric vector: the concentrations of cytosolic calcium [nmol].}

\item{dt}{A numeric, the time interval between two output samples.}

\item{vol}{A numeric, the volume of the system [l].}

\item{a}{A numeric, the first parameter of the cubic function calculating k_AA.}

\item{b}{A numeric, the second parameter of the cubic function calculating k_AA.}

\item{c}{A numeric, the third parameter of the cubic function calculating k_AA.}

\item{k_IB}{A numeric, the rate of association of CaM to a non-phosphorylated subunit.}

\item{k_BI}{A numeric, the rate of dissociation of CaM from a non-phosphorylated subunit.}

\item{k_PT}{A numeric, the ate of dissociation of Ca from CaM bound to a phosphorylated subunit.}

\item{k_TP}{A numeric, the rate of association of Ca to CaM bound to a phosphorylated subunit.}

\item{k_TA}{A numeric, the rate of dissociation of CaM from a phosphorylated subunit.}

\item{k_AT}{A numeric, the rate of association of CaM to a phosphorylated subunit.}

\item{k_AA}{A numeric, the phenomenological rate constant for autophosphorylation.}

\item{c_I}{A numeric, the coefficient of kinase activity of inactive subunits.}

\item{c_B}{A numeric, the coefficient of kinase activity of bound subunits.}

\item{c_P}{A numeric, the coefficient of kinase activity of phosphorylated subunits.}

\item{c_T}{A numeric, the coefficient of kinase activity of trapped subunits.}

\item{c_A}{A numeric, the coefficient of kinase activity of autonomous subunits.}

\item{camT}{A numeric, the total concentration of calmodulin [nmol].}

\item{Kd}{A numeric, the half maximal concentration characterizing Ca binding to CaM.}

\item{Vm_phos}{A numeric, the maximal velocity of the phosphatase reactions.}

\item{Kd_phos}{A numeric, the half maximal concentration characterizing phosphorylated subunits being dephosphorylated.}

\item{totalC}{A numeric, the total fraction of subunits that can phosphorylate or be phosphorylated.}

\item{Wi_conc}{A numeric, the initial concentration of the inactive fraction [nmol].}

\item{Wb_conc}{A numeric, the initial concentration of the bound fraction [nmol].}

\item{Wp_conc}{A numeric, the initial concentration of the phosphorylated fraction [nmol].}

\item{Wt_conc}{A numeric, the initial concentration of the trapped fraction [nmol].}

\item{Wa_conc}{A numeric, the initial concentration of the autonomous fraction [nmol].}

\item{h}{An integer, the Hill coefficient of Ca binding.}
}
\value{
A dataframe with time and the active protein time series as columns.
}
\description{
Takes a calcium time series and simulates the Ca-dependent protein CamKII.
}
\details{
Implementation based on the CamKII Model by Dupont (2003, 2010)
}
\examples{
camkii()
}
