% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{camkii_calculate_amu}
\alias{camkii_calculate_amu}
\title{Propensity Calculation}
\usage{
camkii_calculate_amu()
}
\value{
None
}
\description{
Calculates the propensities of all CamKII model reactions and stores them in the vector amu.
}
\examples{
calculate_amu() 
}
