% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{camkii_props}
\alias{camkii_props}
\title{Calculate propensity for a Camkii Model reaction.}
\usage{
camkii_props(part_num, calcium, rId)
}
\arguments{
\item{part_num}{A numeric vector: the particle numbers of the model species.}

\item{calcium}{A numeric vector: the calcium particle number.}

\item{rId}{An integer value: the id of the specified reaction for which the propensity should be calculated.}
}
\value{
A double value (the propensity of the specified reaction).
}
\description{
Return the propensity of a Camkii Model reaction for a given vector of particle numbers and a reaction Id.
}
\examples{
camkii_props()
}
