% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{camkii_stM}
\alias{camkii_stM}
\title{Define stoichiometric matrix of the CamKII model}
\usage{
camkii_stM()
}
\value{
A numeric matrix: the stoichiometric matrix
}
\description{
Create and return the stoichiometric matrix of the CamKII model as a numeric matrix.
}
\examples{
camkii_stM()
}
