% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcineurin_ode_det.R
\name{detSim_calcineurin}
\alias{detSim_calcineurin}
\title{Specific wrapper function for calcineurin calling the LSODA simulation function (provided by package deSolve)}
\usage{
detSim_calcineurin(input_df, input_sim_params, input_model_params)
}
\arguments{
\item{input_df}{A Dataframe: the input Calcium time series (with at least two columns: "time" in s and "Ca" in nMol/l).}

\item{input_sim_params}{A NumericVector: contains values for the simulation end ("endTime") and its timesteps ("timestep").}

\item{input_model_params}{A List: the model specific parameters. Can contain up to three different vectors named "vols" (model volumes), "init_conc" (initial conditions) and "params" (propensity equation parameters).}
}
\value{
the result of calling the lsoda simulation algorithm from deSolve
}
\description{
Specific wrapper function for calcineurin calling the LSODA simulation function (provided by package deSolve)
}
\examples{
detSim_calcineurin()
}
