% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmodulin_ode_det.R
\name{detSim_calmodulin}
\alias{detSim_calmodulin}
\title{Specific wrapper function for calmodulin calling the LSODA simulation function (provided by package deSolve)}
\usage{
detSim_calmodulin(input_df, sim_params, input_model_params)
}
\arguments{
\item{user_input_df}{A Dataframe: the input Calcium time series (with at least two columns: "time" in s and "Ca" in nMol/l).}

\item{user_sim_params}{A NumericVector: contains values for the simulation end ("endTime") and its timesteps ("timestep").}

\item{user_model_params}{A List: the model specific parameters. Can contain up to three different vectors named "vols" (model volumes), "init_conc" (initial conditions) and "params" (propensity equation parameters).}
}
\value{
the result of calling the lsoda simulation algorithm from deSolve
}
\description{
Specific wrapper function for calmodulin calling the LSODA simulation function (provided by package deSolve)
}
\examples{
detSim_calmodulin()
}
