% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pkc_calculate_amu}
\alias{pkc_calculate_amu}
\title{Propensity Calculation}
\usage{
pkc_calculate_amu()
}
\value{
None
}
\description{
Calculates the propensities of all PKC model reactions and stores them in the vector amu.
}
\examples{
calculate_amu() 
}
