% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_ano}
\alias{sim_ano}
\title{Ano1 Model R Wrapper Function (exported to R)}
\usage{
sim_ano(user_input_df, user_sim_params, user_model_params)
}
\description{
This function calls the internal C++ simulator function to simulate the Ano1 model.
}
\examples{
sim_ano()
}
