% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_camkii}
\alias{sim_camkii}
\title{CamKII Model R Wrapper Function (exported to R)}
\usage{
sim_camkii(user_input_df, user_sim_params, user_model_params)
}
\description{
This function calls the internal C++ simulator function to simulate the CamKII model.
}
\examples{
sim_camkii()
}
