% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_glycphos}
\alias{sim_glycphos}
\title{Glycogen phosphorylase Model R Wrapper Function (exported to R)}
\usage{
sim_glycphos(user_input_df, user_sim_params, user_model_params)
}
\description{
This function calls the internal C++ simulator function to simulate the glycogen phosphorylase model (Gall 2000).
}
\examples{
sim_glycphos()
}
