% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pkc}
\alias{sim_pkc}
\title{PKC Model Wrapper Function (exported to R)}
\usage{
sim_pkc(param_time, param_calcium, param_timestep, param_vol, param_init_conc)
}
\description{
This function calls the internal C++ simulator function to simulate the PKC model.
}
\examples{
model_pkc()
}
