% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulator}
\alias{simulator}
\title{Couple a simulated Ca-dependent protein to a given calcium time series.}
\usage{
simulator(param_time, param_calcium, param_timestep, param_vol, param_init_conc)
}
\arguments{
\item{param_time}{A numeric vector: the times of the observations.}

\item{param_calcium}{A numeric vector: the concentrations of cytosolic calcium [nmol/l].}

\item{param_timestep}{A numeric: the time interval between two output samples.}

\item{param_vol}{A numeric: the volume of the system [l].}

\item{param_init_conc}{A numeric vector: the initial concentrations of model species [nmol/l].}
}
\value{
A dataframe with time and the active protein time series as columns.
}
\description{
Takes a calcium time series and simulates the coupled Ca-dependent protein.
}
\examples{
simulator()
}
