% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulator2}
\alias{simulator2}
\title{Couple a simulated PKC protein to a given calcium time series.}
\usage{
simulator2(param_time, param_calcium, param_timestep, param_vol,
  param_init_conc, calc_propensities, update_system_state)
}
\arguments{
\item{param_time}{A numeric vector: the times of the observations.}

\item{param_calcium}{A numeric vector: the concentrations of cytosolic calcium [nmol/l].}

\item{param_timestep}{A double, the time interval between two output samples.}

\item{param_vol}{A double, the volume of the system [l].}

\item{param_init_conc}{A numeric vector: the initial concentrations of the model species [nmol/l].}

\item{calculate_amu}{A function: provides parameter values and result of propensity equations.}

\item{update_system}{A function: contains stoichiometric effects of all reactions.}
}
\value{
A dataframe with time and the active protein time series as columns.
}
\description{
Takes a calcium time series and simulates the Ca-dependent protein PKC.
}
\details{
Implementation based on the PKC Model by Manninnen (2006)
}
\examples{
simulator2()
}
