% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_system_state}
\alias{update_system_state}
\title{System Update}
\usage{
update_system_state(rIndex)
}
\arguments{
\item{rIndex}{An unsigned integer: the id of the chosen reaction.}
}
\value{
void
}
\description{
Changes the system state (updates the particle numbers) by instantiating a chosen reaction.
}
\examples{
update_system_state() 
}
