% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calmodulin_props}
\alias{calmodulin_props}
\title{Calculate propensity for a Calmodulin Model reaction.}
\usage{
calmodulin_props(x, calcium, rId)
}
\arguments{
\item{x}{A numeric vector: the particle numbers of the model species.}

\item{calcium}{A numeric vector: the calcium particle number.}

\item{rId}{An integer value: the id of the specified reaction for which the propensity should be calculated.}
}
\value{
A double value (the propensity of the specified reaction).
}
\description{
Return the propensity of a Calmodulin Model reaction for a given vector of particle numbers and a reaction Id.
}
\examples{
calmodulin_props()
}
