% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{camkii_props}
\alias{camkii_props}
\title{Calculate propensities for the CamKII Model.}
\usage{
camkii_props(part_num, calcium)
}
\arguments{
\item{part_num}{A numeric vector: the particle numbers of the model species.}

\item{calcium}{A numeric vector: the calcium particle number.}

\item{rId}{An integer value: the id of the specified reaction for which the propensity should be calculated.}
}
\value{
A numeric vector containing a cumulative sum of all reaction propensities.
}
\description{
Return the propensity vector of the CamKII Model for a given vector of particle numbers.
}
\examples{
camkii_props()
}
