% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_ano}
\alias{sim_ano}
\title{Ano1 Model R Wrapper Function (exported to R)}
\usage{
sim_ano(user_input_df, user_sim_params, user_model_params)
}
\arguments{
\item{user_input_df}{A Dataframe: the input Calcium time series (with at least two columns: "time" in s and "Ca" in nmol/l).}

\item{user_sim_params}{A NumericVector: contains values for the simulation end ("endTime") and its timesteps ("timestep").}

\item{user_model_params}{A List: the model specific parameters. Can contain up to three different vectors named "vols" (model volumes), "init_conc" (initial conditions) and "params" (propensity equation parameters).}
}
\value{
the result of calling the model specific version of the function "simulator"
}
\description{
This function compares user-supplied parameters to defaults parameter values, overwrites the defaults if neccessary, and calls the internal C++ simulation function for the ano model.
}
\section{Default Parameters of the Ano1 Model}{

Default Volumes:
\itemize{
\item vol = 1e-11
}

Default Initial Conditions:
\itemize{
\item Cl_ext = 300
\item C = 100
\item C_c = 0
\item C_1 = 0
\item C_1c = 0
\item C_2 = 0
\item C_2c = 0
\item O = 0
\item O_c = 0
\item O_1 = 0
\item O_1c = 0
\item O_2 = 0
\item O_2c = 0
}

Default Reaction Parameters:
\itemize{
\item Vm = -0.06
\item T = 293.15
\item a1 = 0.0077
\item b1 = 917.1288
\item k01 = 0.5979439
\item k02 = 2.853
\item acl1 = 1.8872
\item bcl1 = 5955.783
\item kccl1 = 1.143e-12
\item kccl2 = 0.0009
\item kocl1 = 1.1947e-06
\item kocl2 = 3.4987
\item za1 = 0
\item zb1 = 0.0064
\item zk01 = 0
\item zk02 = 0.1684
\item zacl1 = 0.1111
\item zbcl1 = 0.3291
\item zkccl1 = 0.1986
\item zkccl2 = 0.0427
\item zkocl1 = 0.6485
\item zkocl2 = 0.03
\item l = 41.6411
\item L = 0.6485
\item m = 0.0102
\item M = 0.0632
\item h = 0.3367
\item H = 14.2956
}
}

\examples{
sim_ano()
}
