% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_calcineurin}
\alias{sim_calcineurin}
\title{Calcineurin Model R Wrapper Function (exported to R)}
\usage{
sim_calcineurin(user_input_df, user_sim_params, user_model_params)
}
\description{
This function calls the internal C++ simulator function to simulate the Calcineurin model (Fisher 2006).
}
\examples{
sim_calcineurin()
}
