% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_calmodulin}
\alias{sim_calmodulin}
\title{Calmodulin Model R Wrapper Function (exported to R)}
\usage{
sim_calmodulin(user_input_df, user_sim_params, user_model_params)
}
\description{
This function calls the internal C++ simulator function to simulate the Calmodulin model.
}
\examples{
sim_calmodulin()
}
