% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pkc}
\alias{sim_pkc}
\title{PKC Model R Wrapper Function (exported to R)}
\usage{
sim_pkc(user_input_df, user_sim_params, user_model_params)
}
\description{
This function calls the internal C++ simulator function to simulate the PKC model.
}
\examples{
sim_pkc()
}
