% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulator}
\alias{simulator}
\title{Stochastic Simulatior (Gillespie).}
\usage{
simulator(time, calcium, init_conc, calc_props, provide_stM, dt, vol)
}
\arguments{
\item{time}{A numeric vector: the times of the observations.}

\item{calcium}{A numeric vector: the particle numbers of cytosolic calcium [nmol].}

\item{init_conc}{A numeric vector: the initial concentrations of the model species.}

\item{calc_props}{A function: returns a vector of model propensities for a given vector of concentrations.}

\item{provide_stM}{A function: returns a matrix of the stoichiometric coefficients of the reaction system.}

\item{dt}{A numeric, the time interval between two output samples.}

\item{vol}{A numeric, the volume of the system [l].}
}
\value{
A dataframe with time and the active protein time series as columns.
}
\description{
Simulate a model using an implementation of Gillespie's SSA.
}
\examples{
simulator()
}
